+++ apps/wallet/src/components/inputs/StandardsAutocomplete.vue
@@ -1,27 +1,27 @@
 <template>
   <VCombobox
     v-model="model"
-    :multiple="props.multiple.value"
-    :label="props.label.value"
+    :multiple="props.multiple"
+    :label="props.label"
     item-value="value"
     item-title="text"
     :items="items"
-    :variant="props.variant.value"
-    :density="props.density.value"
-    :readonly="props.readonly.value"
-    :disabled="props.disabled.value"
-    :rules="props.rules.value"
+    :variant="props.variant"
+    :density="props.density"
+    :readonly="props.readonly"
+    :disabled="props.disabled"
+    :rules="props.rules"
   />
 </template>
 
 <script setup lang="ts">
-import { computed, toRefs } from 'vue';
+import { computed } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { VCombobox } from 'vuetify/components';
 import { useStationStore } from '~/stores/station.store';
 import { FormValidationRuleFn } from '~/types/helper.types';
 
-const input = withDefaults(
+const props = withDefaults(
   defineProps<{
     modelValue?: string[];
     blockchain: string;
@@ -45,18 +45,17 @@ const input = withDefaults(
   },
 );
 
-const props = toRefs(input);
+// const props = toRefs(input);
 const i18n = useI18n();
 const station = useStationStore();
 const standardsData = computed(
   () =>
-    station.configuration.details.supported_blockchains.find(
-      b => b.blockchain === props.blockchain.value,
-    )?.supported_standards || [],
+    station.configuration.details.supported_blockchains.find(b => b.blockchain === props.blockchain)
+      ?.supported_standards || [],
 );
 
 const model = computed({
-  get: () => props.modelValue.value,
+  get: () => props.modelValue,
   set: value => {
     const standards = value as { value: string; text: string }[] | undefined;
 
@@ -74,7 +73,7 @@ const emit = defineEmits<{
 const items = computed(() =>
   standardsData.value.map(data => ({
     value: data.standard,
-    text: i18n.t(`blockchains.${props.blockchain.value}.standards.${data.standard}`),
+    text: i18n.t(`blockchains.${props.blockchain}.standards.${data.standard}`),
   })),
 );
 </script>
